backend/app/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
geoalchemy2==0.14.7
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
redis==5.0.4
numpy==1.26.4
folium==0.16.0
alembic==1.13.1
aioredis==2.0.1
gpxpy==1.6.2
geopandas==0.14.4
shapely==2.0.4
pandas==2.2.2
scipy==1.13.0
plotly==5.22.0
jinja2==3.1.4
python-dotenv==1.0.1
bcrypt==4.0.1